refactor(trusty-common): extract pure palace-ID derivation from trusty-memory (refs #1605)
WI-1: move the pure palace-slug derivation core out of
crates/trusty-memory/src/palace_id_derive.rs into a new
crates/trusty-common/src/palace_id.rs so trusty-mpm can derive the
identical palace slug for managed-session MCP injection without a
trusty-mpm -> trusty-memory dependency edge (mirrors how derive_index_id
is shared for trusty-search index pinning, #1373).

- New trusty-common palace_id module with derive_palace_id,
  owner_repo_from_git_remote, parent_dir_slug, PALACE_OVERRIDE_ENV,
  palace_override_from_env + all precedence/URL-variant unit tests.
- Re-exported at the trusty-common crate root.
- trusty-memory palace_id_derive.rs is now a thin re-export shim so every
  existing call site (messaging::operations, messaging::mod) keeps working
  unchanged. ZERO behavior change.
